requirements.txt
PyMuPDF==1.27.2
rapidocr-onnxruntime==1.2.3
opencv-python==4.12.0.88
numpy==2.1.3
Pillow==11.1.0
lxml==6.0.1
PyYAML==6.0.2
reportlab==4.4.4
pytest==9.0.2
requests==2.32.5
langchain==1.2.13
langchain-core==1.2.20
langchain-openai==1.1.11
opendataloader-pdf==2.1.1
